cleanup: move hook code
